% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMmboost.R
\name{ELMmboost}
\alias{ELMmboost}
\title{SurvELM ELMmboost}
\usage{
ELMmboost(x, y, Kernel_type = "lin_kernel", Kernel_para = c(2, 1), ...)
}
\arguments{
\item{x}{The covariates(predictor variables) of training data.}

\item{y}{Survival time and censored status of training data. Must be a Surv  \code{survival} object}

\item{Kernel_type}{Type of kernel matrix. Currently four options avaibable. "RBF_kernel",a RBF kernel;"lin_kernel" , a linear kernel;poly_kernel ,a polynomial kernel;sigmoid_kernel, a sigmoid kernel. Default is "lin_kernel".}

\item{Kernel_para}{Parameters for different types of kernels. A single value for RBF and linear kernels. A vector for polynomial and sigmoid kernels and progam stops if only a single value is supplied. However, if the vector of values is supplied in the cases of RBF and liner kernels, only the first value will be used. Default is a vector value "c(2,1)".}

\item{...}{Additional arguments for  mboost.}
}
\value{
Object of class \code{ELMmboost} with elements
  \tabular{ll}{
      \code{elmglmboost}    \tab  A glmboost model. See \code{mboost} for details. \cr
      \code{trainx} \tab  Training data covariates. \cr
         \code{kerneltype} \tab  Type of kernel matrix used in training. kerneltype=1,a RBF kernel;kerneltype=2 , a linear kernel;kerneltype=3 ,a polynomial kernel;kerneltype=4, a sigmoid kernel. \cr
  \code{Kernel_para} \tab  Parameters used in training. A single value for kerneltype=1 or 2. A vector for kerneltype=3 or 4. \cr
  }
}
\description{
An Extreme Learning Machine Cox  Model with Gradient Based Boosting
}
\examples{
set.seed(123)
require(SurvELM)
require(survival)
#Lung DATA
data(lung)
lung=na.omit(lung)
lung[,3]=lung[,3]-1
n=dim(lung)[1]
L=sample(1:n,ceiling(n*0.5))
trset<-lung[L,]
teset<-lung[-L,]
rii=c(2,3)
elmsurvmodel=ELMmboost(x=trset[,-rii],y=Surv(trset[,rii[1]], trset[,rii[2]]))
#THE predicted linear predictor
testpre=predict(elmsurvmodel,teset[,-c(rii)])
}
\references{
\itemize{
  \item Hong Wang et al (2017). A Survival Ensemble of Extreme Learning Machine. Applied Intelligence, DOI:10.1007/s10489-017-1063-4.
 }
}
\seealso{
\code{\link{mboost}}
}
\author{
Hong Wang
}
