% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ELMBJEN.R
\name{predict.ELMBJEN}
\alias{predict.ELMBJEN}
\title{SurvELM predict.ELMBJEN}
\usage{
\method{predict}{ELMBJEN}(object, testx, trlength, ...)
}
\arguments{
\item{object}{An object that inherits from class ELMBJEN.}

\item{testx}{A data frame in which to look for variables with which to predict.}

\item{trlength}{Number of based models used for prediction, shouble be less than and equal to the number for training.}

\item{...}{Additional arguments.}
}
\value{
produces a vector of predictions or a matrix of predictions
}
\description{
Predicting from A Kernel Extreme Learning Machine Ensemble Using the Buckley-James estimator
}
\examples{
set.seed(123)
require(SurvELM)
require(survival)
#Lung DATA
data(lung)
lung=na.omit(lung)
lung[,3]=lung[,3]-1
n=dim(lung)[1]
L=sample(1:n,ceiling(n*0.5))
trset<-lung[L,]
teset<-lung[-L,]
rii=c(2,3)
# with RBF kernel
elmsurvmodel=ELMBJEN(x=trset[,-rii],y=Surv(trset[,rii[1]],trset[,rii[2]]),
trlength=10,Kernel_type="RBF_kernel",Kernel_para=c(2,1))
#The second base model
fit2=elmsurvmodel$elmcoxfit[[2]]
#The predicted survival
testprelin=predict(elmsurvmodel,teset[,-c(rii)])
}
\references{
\itemize{
  \item Hong Wang et al (2017). A Survival Ensemble of Extreme Learning Machine. Applied Intelligence, DOI:10.1007/s10489-017-1063-4.
 }
}
\seealso{
\code{predict.ELMBJ}
}
\author{
Hong Wang
}
