% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ELMCox.R
\name{predict.ELMCox}
\alias{predict.ELMCox}
\title{SurvELM predict.ELMCox}
\usage{
\method{predict}{ELMCox}(object, testx, ...)
}
\arguments{
\item{object}{An object that inherits from class ELMCox.}

\item{testx}{A data frame in which to look for variables with which to predict.}

\item{...}{Additional arguments for  \code{glmnet}.}
}
\value{
produces a vector of predictions or a matrix of predictions
}
\description{
Predicting from A Regularized Cox  Extreme Learning Machine Model
}
\examples{
set.seed(123)
require(SurvELM)
require(survival)
#Lung DATA
data(lung)
lung=na.omit(lung)
lung[,3]=lung[,3]-1
n=dim(lung)[1]
L=sample(1:n,ceiling(n*0.5))
trset<-lung[L,]
teset<-lung[-L,]
rii=c(2,3)
# Default with lasso penalty
elmsurvmodel=ELMCox(x=trset[,-rii],y=Surv(trset[,rii[1]], trset[,rii[2]]))
# with ridge penalty and RBF kernel, alpha has the same meaning as in glmnet
elmsurvmodel=ELMCox(x=trset[,-rii],y=Surv(trset[,rii[1]], 
trset[,rii[2]]),Kernel_type="RBF_kernel",Kernel_para=c(2,1),alpha=0)
# with elastic net penalty
elmsurvmodel=ELMCox(x=trset[,-rii],y=Surv(trset[,rii[1]], trset[,rii[2]]),alpha=0.5)
#The predicted linear predictor
testprelin=predict(elmsurvmodel,teset[,-c(rii)],type="link")
#The predicted  relative-risk
testpreres=predict(elmsurvmodel,teset[,-c(rii)],type="response")
}
\seealso{
\code{predict.glmnet}
}
\author{
Hong Wang
}
