% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ELMCoxBoost.R
\name{predict.ELMCoxBoost}
\alias{predict.ELMCoxBoost}
\title{SurvELM predict.ELMCoxBoost}
\usage{
\method{predict}{ELMCoxBoost}(object, testx, ...)
}
\arguments{
\item{object}{An object that inherits from class ELMCoxBoost.}

\item{testx}{A data frame in which to look for variables with which to predict.}

\item{...}{Additional arguments for  \code{CoxBoost}.}
}
\value{
produces a vector of predictions or a matrix of predictions
}
\description{
Predicting from An Extreme Learning Machine Cox  Model with Likelihood Based Boosting
}
\examples{
set.seed(123)
library(SurvELM)
library(survival)
#Lung DATA
data(lung)
lung=na.omit(lung)
lung[,3]=lung[,3]-1
n=dim(lung)[1]
L=sample(1:n,ceiling(n*0.5))
trset<-lung[L,]
teset<-lung[-L,]
rii=c(2,3)
elmsurvmodel=ELMCoxBoost(x=trset[,-rii],y=Surv(trset[,rii[1]], trset[,rii[2]]))
#THE predicted linear predictor
testpre=predict(elmsurvmodel,teset[,-c(rii)])
#The predicted cumulative incidence function
testprecif=predict(elmsurvmodel,teset[,-c(rii)],type="CIF")
# The predicted partial log-likelihood
testprellk=predict(elmsurvmodel,teset[,-c(rii)],newtime=teset[,rii[1]],
newstatus=teset[,rii[2]],type="logplik")
uniquetimes=sort(unique(trset$time))
# The predicted probability of not yet having had the event at the time points given in times
testprerisk=predict(elmsurvmodel,teset[,-c(rii)],times=uniquetimes,type="risk")
}
\seealso{
\code{predict.CoxBoost}
}
\author{
Hong Wang
}
