% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ELMCoxEN.R
\name{predict.ELMCoxEN}
\alias{predict.ELMCoxEN}
\title{SurvELM predict.ELMCoxEN}
\usage{
\method{predict}{ELMCoxEN}(object, testx, enlen, ...)
}
\arguments{
\item{object}{An object that inherits from class ELMCoxEN.}

\item{testx}{A data frame in which to look for variables with which to predict.}

\item{enlen}{Number of based models used for prediction, shouble be less than and equal to the number for training.}

\item{...}{Additional arguments for  \code{glmnet}.}
}
\value{
produces a vector of predictions or a matrix of predictions
}
\description{
Predicting from an Ensemble of Regularized Cox  Extreme Learning Machine  Model
}
\examples{
set.seed(123)
require(SurvELM)
require(survival)
#Lung DATA
data(lung)
lung=na.omit(lung)
lung[,3]=lung[,3]-1
n=dim(lung)[1]
L=sample(1:n,ceiling(n*0.5))
trset<-lung[L,]
teset<-lung[-L,]
rii=c(2,3)
# with ridge penalty and RBF kernel, alpha has the same meaning as in glmnet
elmsurvmodel=ELMCoxEN(x=trset[,-rii],y=Surv(trset[,rii[1]],trset[,rii[2]]),
enlen=10,Kernel_type="RBF_kernel",Kernel_para=c(2,1),alpha=0)
#The second base model
fit2=elmsurvmodel$elmcoxfit[[2]]
#The predicted linear predictor
testprelin=predict(elmsurvmodel,teset[,-c(rii)],type="link")
#The predicted  relative-risk
testpreres=predict(elmsurvmodel,teset[,-c(rii)],type="response")
}
\seealso{
\code{predict.glmnet}
}
\author{
Hong Wang
}
