% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ELMmboost.R
\name{predict.ELMmboost}
\alias{predict.ELMmboost}
\title{SurvELM predict.ELMmboost}
\usage{
\method{predict}{ELMmboost}(object, testx, ...)
}
\arguments{
\item{object}{An object that inherits from class ELMmboost.}

\item{testx}{A data frame in which to look for variables with which to predict.}

\item{...}{Additional arguments for  mboost.}
}
\value{
produces a vector of predictions or a matrix of predictions
}
\description{
Predicting from an Extreme Learning Machine Cox  Model with Gradient Based Boosting
}
\examples{
set.seed(123)
require(SurvELM)
require(survival)
#Lung DATA
data(lung)
lung=na.omit(lung)
lung[,3]=lung[,3]-1
n=dim(lung)[1]
L=sample(1:n,ceiling(n*0.5))
trset<-lung[L,]
teset<-lung[-L,]
rii=c(2,3)
elmsurvmodel=ELMmboost(x=trset[,-rii],y=Surv(trset[,rii[1]], trset[,rii[2]]))
#THE predicted linear predictor
testpre=predict(elmsurvmodel,teset[,-c(rii)])
}
\seealso{
\code{\link{mboost}}
}
\author{
Hong Wang
}
