% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survFit.ELMmboost.R
\name{survFit.ELMmboost}
\alias{survFit.ELMmboost}
\title{SurvELM SurvFit.ELMmboost}
\usage{
\method{survFit}{ELMmboost}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An ELMmboost object.}

\item{newdata}{An optional data frame in which to look for variables with which to predict the survivor function.}

\item{...}{Additional arguments for  mboost.}
}
\value{
Object of class \code{survFit} with elements
  \tabular{ll}{
      \code{surv}    \tab  Estimated survival probabilities at the given time points\cr
      \code{time} \tab  The evaluated given time points. \cr
      \code{n.event} \tab  Number of observed events at given at time points. \cr
  }
}
\description{
An Extreme Learning Machine Cox  Model with Gradient Based Boosting
}
\examples{
set.seed(123)
require(SurvELM)
require(survival)
#Lung DATA
data(lung)
lung=na.omit(lung)
lung[,3]=lung[,3]-1
n=dim(lung)[1]
L=sample(1:n,ceiling(n*0.5))
trset<-lung[L,]
teset<-lung[-L,]
rii=c(2,3)
elmsurvmodel=ELMmboost(x=trset[,-rii],y=Surv(trset[,rii[1]], trset[,rii[2]]))
sfit=survFit(elmsurvmodel)
# plot the survival probability
plot(sfit, xlab = "Time", ylab = "Survival Probability")
}
\seealso{
\code{survFit}
}
